Update documentation regarding Kali branches
Make it explicit that kali-rolling and kali-last-snapshot are full
distributions, and user should use either one of those, but not both.

Make it explicit that kali-experimental and kali-bleeding-edge are
partial distributions, meant to be used in addition to kali-rolling.

Add a mention of kali-tweaks, as the prefered way to switch branches.

Nitpick: consistenly order components 'main contrib non-free', as this
is how it usually appears pretty much everywhere. It's what's in a Kali
live image and in a Kali container. Not sure about a Kali install, I
just guess it also appears in this order.