Add conv=fsync to dd command lines
In most of the documentation, we use dd to write Kali images to block
devices.

There's a bit of a grey area regarding the question: is the data
physically written to the block device after dd returns? Or are there
some risk that the data is in a cache somewhere, not yet flushed to the
disk?

There's no satisfying answer to this question. Everybody in the team
seem to agree that running `sync` after dd is good practice, and that
there was a time when it was definitely needed. But we're not quite sure
if it's still needed nowadays. Maybe some it's now done automagically
thanks to changes in the kernel? Or in dd itself? Who knows...

It's also possible that the answer depends on the flash device, maybe
the kernel will cache data for some devices, and won't for others, who
knows...

The most recent and decent blog post I could find on the matter is
https://abbbi.github.io/dd/

Anyway. So we're not 100% sure.

Taking another perspective, we can ask: what happens if someone flash
Kali on a USB drive, unplug it without running sync, and it turns out
that some data is missing on the drive? Not quite sure of the answer
once again, but user might find themselves with either an unbootable
device, or worst, a Kali system that boots and seems to work, until it
fails at some mysterious moment. This is when user file some mysterious
bug reports, with the weirdest failure, and we all scratch our heads
wondering what's wrong.

So! In an attempt to prevent this possible failure, to save our users
some pain, and to save ourselves from having to deal with mysterious bug
reports that nobody quite understands (and let alone reproduce), let's
document the use of `conv=fsync` in all the dd commands.

The documentation for `conf=fdatasync` and `conv=fsync` goes like this:

    fdatasync   physically write output file data before finishing
    fsync       likewise, but also write metadata

It really can't hurt to add it to the command-line. At worst, it won't
have any effect.

Note some possible alternatives:

* run `sync` after `dd`, however running sync doesn't just sync what's
  being written to the sdcard, it actually syncs all of them.
* stop using dd, use `cp ISO /dev/sdX` or `cat ISO > /dev/sdX`, however
  we probably want to run `sync` afterwards anyway (once again, "just in
  case"), so it doesn't really address the problem.