apt documents full-upgrade, not dist-upgrade
Both commands do the same, and both are supported by apt and apt-get.

However, `apt -h` and `man apt` document the command `full-upgrade`, and
do not mention `dist-upgrade`. While apt-get traditionally used the
command `dist-upgrade`, and it's what you see in the documentation.

I think that the background story is that `dist-upgrade` is a bit of a
confusing name (nobody really knows what it means), while `full-upgrade`
carries the meaning better. So `dist-upgrade` is being phased out in
favor of `full-upgrade`, or at least, that's the term that `apt` chosed
to use.

Some ref: https://askubuntu.com/questions/770135/apt-full-upgrade-versus-apt-get-dist-upgrade#comment1899278_770140