advanced packaging: rewrite the first half to use uscan and get the version string right
There are many ways to start a package. In case of a Git snapshot, the
main difficulty is to get the package version right, and this is much
better to let tools figure that out automatically. Uscan does a great
job at that, all we need is to create the `debian/watch` file first.
Once this is done, uscan will download the git snapshot, pack it, and
give it the right version string.

Then it's easy, from this point, to follow the same course of action as
what's documented already in the intro/intermediate packaging, ie. do
`gbp import-orig` and then `dh_make`.