Update the no-sound page, remove mention of the bugtracker
Minor rewords, and remove mention of the bugtracker. Generally we don't
hint at opening a bug in the doc, as there would be too many pages
where we'd say "open a bug if X or Y doesn't work".

In that case I made an exception as we were just transitioning to
pipewire, and I wanted to be sure to get as many bug reports as
possible, to fix whatever might have been broken.

But it's been a year now, and the topic is long behind us.